add nav pages
