+++ src/Components/Header.jsx
@@ -1,0 +1,65 @@
+import React from "react";
+import { Link } from "react-router-dom";
+function Header() {
+  return (
+    <>
+
+      <div class="sub-header">
+        <div class="container">
+          <div class="row">
+            <div class="col-lg-8 col-md-8">
+              <ul class="info">
+                <li><i class="fa fa-envelope"></i> info@company.com</li>
+                <li><i class="fa fa-map"></i> Sunny Isles Beach, FL 33160</li>
+              </ul>
+            </div>
+            <div class="col-lg-4 col-md-4">
+              <ul class="social-links">
+                <li><a href="#"><i class="fab fa-facebook"></i></a></li>
+                <li><a href="https://x.com/minthu" target="_blank"><i class="fab fa-twitter"></i></a></li>
+                <li><a href="#"><i class="fab fa-linkedin"></i></a></li>
+                <li><a href="#"><i class="fab fa-instagram"></i></a></li>
+              </ul>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      {/* <!-- ***** Header Area Start ***** --> */}
+      <header class="header-area header-sticky">
+        <div class="container">
+          <div class="row">
+            <div class="col-12">
+              <nav class="main-nav">
+                {/* <!-- ***** Logo Start ***** --> */}
+                <a href="/" class="logo">
+                  <h1>Villa</h1>
+                </a>
+                {/* <!-- ***** Logo End ***** --> */}
+                {/* <!-- ***** Menu Start ***** --> */}
+                <ul class="nav">
+                  <React.StrictMode>
+                    <browserRouter>
+                      <li><Link to="/" class="active">Home</Link></li>
+                      <li><Link to="/aboutus">Properties</Link></li>
+                      <li><Link to="/contact">Contact Us</Link></li>
+                      <li><Link to="/services">Services</Link></li>
+                      <li><Link to="#"><i class="fa fa-calendar"></i> Schedule a visit</Link></li>
+                    </browserRouter>
+                  </React.StrictMode>
+                </ul>
+                <a class='menu-trigger'>
+                  <span>Menu</span>
+                </a>
+                {/* <!-- ***** Menu End ***** --> */}
+              </nav>
+            </div>
+          </div>
+        </div>
+      </header>
+      {/* <!-- ***** Header Area End ***** --> */}
+
+    </>
+  );
+}
+export default Header;
